Refactor: String[] Names를 일급 콜렉션으로 변경 및 하드코딩 데이터 상수화
