+++ src/test/java/ladder/LineTest.java
@@ -7,7 +7,6 @@ public class LineTest {
     @Test
     public void lineTest() {
         Line line = new Line(4);
-        line.makeRow();
         line.printRow();
     }
 }
